Add `goto` command
