+++ src/app/commands/index.go
@@ -24,6 +24,7 @@ type Cli struct {
 
 	// Misc
 	Edit    Edit    `cmd:"" group:"Misc" help:"Opens a file or bookmark in your editor"`
+	Goto    Goto    `cmd:"" group:"Misc" help:"Opens directory of file in your file explorer"`
 	Json    Json    `cmd:"" group:"Misc" help:"Converts records to JSON"`
 	Info    Info    `cmd:"" group:"Misc" default:"withargs" help:"Displays meta info about klog"`
 	Widget  Widget  `cmd:"" group:"Misc" help:"Starts menu bar widget (MacOS only)" hidden:""`
