--- src/app/lib/terminalformat/init_windows.go
@@ -1,25 +1,0 @@
-//go:build windows
-
-package terminalformat
-
-import (
-	"os"
-	"syscall"
-	"unsafe"
-)
-
-const enableVirtualTerminalProcessing = 0x0004
-
-var (
-	kernel32           = syscall.NewLazyDLL("kernel32.dll")
-	procGetConsoleMode = kernel32.NewProc("GetConsoleMode")
-	procSetConsoleMode = kernel32.NewProc("SetConsoleMode")
-)
-
-func init() {
-	var mode uint32
-	procGetConsoleMode.Call(os.Stdout.Fd(), uintptr(unsafe.Pointer(&mode)))
-	if (mode & enableVirtualTerminalProcessing) != enableVirtualTerminalProcessing {
-		procSetConsoleMode.Call(os.Stdout.Fd(), uintptr(mode|enableVirtualTerminalProcessing))
-	}
-}
